Enum telerunbot::bot::Command

enum Command {
    Help,
    Show,
    Add {
        distance: f32,
        user_name: String,
    },
    Edit {
        run_id: i32,
        distance: f32,
    },
    Delete {
        run_id: i32,
    },
    Tally,
    List {
        limit: u32,
    },
}

Enumeration of commands accepted by the bot.

Variants

Help

Matched to /help -> displays commands and their documentation.

Show

Matched to /show -> displays users within chat.

Add

Fields

distance: f32

Distance run in km

user_name: String

Name of user to tie the run to. Must be unique.

Matched to /add <distance> <user_name> -> creates users in db if not present, then adds run data to runs table.

Edit

Fields

run_id: i32

Id of run as stored in runs table.

distance: f32

Corrected distance run in km.

Matched to /edit <run_id> <distance> -> edits stored run data.

Delete

Fields

run_id: i32

Id of run to remove from table.

Matched to /delete <run_id> -> removes a certain run from database.

Tally

Matched to /tally -> sends score board as message through Telegram.

List

Fields

limit: u32

Limit to query from db.

Matched to /list <limit> -> displays runs registered by the group chat, subject to a limit.