Module telerunbot::message

Templating and message preparation.

This module reads in Rust-native objects and renders them as Strings using askama as the templating engine.

Structs

ListRunTemplate 🔒

Struct Run display.

ListTallyTemplate 🔒

Struct Tally display.

ListUserTemplate 🔒

Struct User display.

RunDisplay 🔒

NewType implementation so that Display can be implemented for it.

Functions

display_tally

Displays score aggregates fetched from database.

list_runs

Displays runs fetched from database.

list_users

Displays users fetched from database.